refactor: add serverless router service
